ai-automation-flow-main/server/requirements.txt
flask>=3.1.2
flask-cors>=6.0.2
flask-apscheduler>=1.13.1
openai>=2.14.0
requests>=2.32.5
croniter>=2.0.1
psycopg2-binary>=2.9.9
cryptography>=41.0.7
